junto las llamadas a la API en un solo archivo.
Creo llamadasApi que contendra todas las llamadas a la API.
Importo la funcion que necesito directamente a formulario_turno.js y la llamo al principio.
