+++ html/formulario_turno.html
@@ -77,7 +77,6 @@
 
   <!--  .js  -->
   <script src="../js/formulario_turno.js"></script>
-  <script src="../js/completarEspecialidades.js"></script>
 
 </body>
