+++ js/formulario_turno.js
@@ -1,3 +1,7 @@
+import { cargarEspecialidades } from "./llamadasApi";
+
+cargarEspecialidades();
+
 document.getElementById("especialidad").addEventListener("change", () => {
     especialidad_check()
 });
@@ -8,7 +12,7 @@ document.getElementById("fecha").addEventListener("change", () => {
 
 document.addEventListener("mouseover", event => {
     let idHorario = event.target.id;
-    if(idHorario >= 0 & idHorario<= 15 & idHorario!=""){
+    if (idHorario >= 0 & idHorario <= 15 & idHorario != "") {
         hoverColorearHorarios(idHorario);
     }
 });
@@ -16,45 +20,45 @@ document.addEventListener("mouseover", event => {
 
 document.addEventListener("mouseout", event => {
     let idHorario = event.target.id;
-    if(idHorario >= 0 & idHorario<= 15 & idHorario!=""){
+    if (idHorario >= 0 & idHorario <= 15 & idHorario != "") {
         hoverDescolorearHorarios(idHorario);
     }
 });
 
 document.addEventListener("click", event => {
     let idHorario = event.target.id;
-    if(idHorario >= 0 & idHorario<= 15 & idHorario!=""){
+    if (idHorario >= 0 & idHorario <= 15 & idHorario != "") {
         clickColorearHorarios(idHorario);
     }
 });
 
-function hoverColorearHorarios(idHorario){
+function hoverColorearHorarios(idHorario) {
     let CantidadModulos = document.getElementById("especialidad").value;
     let listaHorarios = document.querySelectorAll(".grid-cell");
     let ultimoHorario = parseInt(idHorario) + parseInt(CantidadModulos);
-    for (let i=idHorario; i<ultimoHorario; i++){
+    for (let i = idHorario; i < ultimoHorario; i++) {
         listaHorarios[i].classList.toggle("bg-warning");
     }
 }
 
-function hoverDescolorearHorarios(idHorario){
+function hoverDescolorearHorarios(idHorario) {
     let CantidadModulos = document.getElementById("especialidad").value;
     let listaHorarios = document.querySelectorAll(".grid-cell");
     let ultimoHorario = parseInt(idHorario) + parseInt(CantidadModulos);
-    for (let i=idHorario; i<ultimoHorario; i++){
+    for (let i = idHorario; i < ultimoHorario; i++) {
         listaHorarios[i].classList.remove("bg-warning");
     }
 }
 
-function clickColorearHorarios(idHorario){
+function clickColorearHorarios(idHorario) {
     //ELIMINA VERDE DE TODOS Y DESPUES APLICA EN LOS SELECCIONADOS
     let CantidadModulos = document.getElementById("especialidad").value;
     let listaHorarios = document.querySelectorAll(".grid-cell");
     let ultimoHorario = parseInt(idHorario) + parseInt(CantidadModulos);
-    for (let i=0; i<listaHorarios.length; i++){
+    for (let i = 0; i < listaHorarios.length; i++) {
         listaHorarios[i].classList.remove("bg-success");
     }
-    for (let i=idHorario; i<ultimoHorario; i++){
+    for (let i = idHorario; i < ultimoHorario; i++) {
         listaHorarios[i].classList.add("bg-success");
     }
     document.getElementById("confirmar").classList.remove("hidden");
@@ -64,7 +68,7 @@ function clickColorearHorarios(idHorario){
 function especialidad_check() {
     let especialidad = document.getElementById("especialidad");
     let fecha = document.getElementById("fecha");
-    if (especialidad.value !="" & fecha.classList.contains("hidden")) {
+    if (especialidad.value != "" & fecha.classList.contains("hidden")) {
         fecha.classList.toggle("hidden");
     }
 }
@@ -72,7 +76,7 @@ function especialidad_check() {
 function fecha_check() {
     let fecha = document.getElementById("fecha");
     let horario = document.getElementById("horario");
-    if (fecha.value !="" & horario.classList.contains("hidden")) {
+    if (fecha.value != "" & horario.classList.contains("hidden")) {
         horario.classList.toggle("hidden");
     }
 }
